+++ bfDescription/bfDescription.html
@@ -1,5 +1,5 @@
-<link rel="stylesheet" href="./bfDescription/styles.css">
-<script type="module" src="./bfDescription/bfDescription.js"></script>
+<link rel="stylesheet" href="/bfDescription/styles.css">
+<script type="module" src="/bfDescription/bfDescription.js"></script>
 <div id="bf-description">
     Self-modifying version of <a href="https://en.wikipedia.org/wiki/Brainfuck" target=”_blank”>Brainfuck</a>
     esoteric programming language. The program modifies itself as it executes. Each instruction
