+++ bfInterpreter/bfInterpreter.html
@@ -1,5 +1,5 @@
-<link rel="stylesheet" href="./bfInterpreter/styles.css">
-<script type="module" src="./bfInterpreter/visualizeBrainfuck.js"></script>
+<link rel="stylesheet" href="/bfInterpreter/styles.css">
+<script type="module" src="/bfInterpreter/visualizeBrainfuck.js"></script>
 <h1>Brainfuck Interpreter</h1>
 <div>
     <div class="controls">
