+++ pvp/board.html
@@ -1,5 +1,5 @@
 <link rel="stylesheet" href="/board/styles.css">
-<script type="module" src="/board/visualizeBoard.js"></script>
+<script type="module" src="/pvp/pvp.js"></script>
 <div>
     <div>
         <h1>Petri Cells</h1>
