+++ pvp/index.html
@@ -5,7 +5,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>petri_cells</title>
-    <link rel="stylesheet" href="styles.css">
+    <link rel="stylesheet" href="/styles.css">
     <script src="/shared/inputValidation.js"></script>
     <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
     <script> 
@@ -21,7 +21,7 @@
 
 <body>
     <div class="section">
-        <div data-include="/board/board"></div>
+        <div data-include="/pvp/board"></div>
     </div>
     <hr class="divider">
     </hr>
