+++ pvp/pvp.js
@@ -1,6 +1,6 @@
 import { EventHandleHelper } from "../shared/handleEvents.js";
-import { GridController } from "./gridController.js"
-import { HistoryManager } from "./historyManager.js";
+import { GridController } from "../board/gridController.js"
+import { HistoryManager } from "../board/historyManager.js";
 
 const HISTORY_FIDELITY = 20;
 const inputtedWidth = document.getElementById("bf-w")[0].value;
@@ -9,7 +9,7 @@ const inputtedHeight = document.getElementById("bf-h")[0].value;
 const height = parseFloat(inputtedHeight);
 const controller = new GridController();
 const store = {
-  state: controller.initState(width, height),
+  state: controller.initStateToData(width, height),
   uiItems: controller.initGridUI(width, height),
 };
 const history = new HistoryManager().init(HISTORY_FIDELITY, store.state);
@@ -45,11 +45,23 @@ eventHandleHelper.addEventListener(buttonMapping.runButton, () => {
 });
 eventHandleHelper.addEventListener(buttonMapping.restartButton, () => {
   const inputtedWidth = document.getElementById("bf-w")[0].value;
-  const width = parseFloat(inputtedWidth) || 20;
+  const width = parseFloat(inputtedWidth) || 10;
   const inputtedHeight = document.getElementById("bf-h")[0].value;
-  const height = parseFloat(inputtedHeight) || 20;
+  const height = parseFloat(inputtedHeight) || 10;
+
+  // Get user input for two programs
+  const program1Input = prompt("Enter the first program:");
+  const program2Input = prompt("Enter the second program:");
+
+  const program1 = controller.logic.fromGenericInput(program1Input);
+  const program2 = controller.logic.fromGenericInput(program2Input);
+
   store.uiItems = controller.initGridUI(width, height);
-  store.state = controller.initState(width, height);
+  store.state = controller.initStateToData(width, height);
+  store.state.grid = controller.placeProgramsRandomly(
+    store.state.grid,
+    [program1, program2],
+  );
   history.init(HISTORY_FIDELITY, store.state);
   controller.updateGridUI(store);
   controller.stopRunning(runButton);
@@ -80,7 +92,7 @@ document
     }
   });
 eventHandleHelper.addEventListener("cell-details-1-edit-submit", () => {
-  const inputVal = document.getElementById("cell-details-1-edit-input").value
+  const inputVal = document.getElementById("cell-details-1-edit-input").value;
   controller.editProgramWithNumsForm(store.state, inputVal);
 });
 eventHandleHelper.addEventListener("cell-details-2-edit-submit", () => {
